Alteração nas associações.

+++ src/main/java/com/marceloproject/webservicespring/entities/Product.java
@@ -13,7 +13,7 @@ public class Product implements Serializable{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private Double price;
@@ -27,7 +27,7 @@ public class Product implements Serializable{
 
     public Product(){}
 
-    public Product(Integer id, String name, String description, Double price, String imgUrl) {
+    public Product(Long id, String name, String description, Double price, String imgUrl) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -35,11 +35,11 @@ public class Product implements Serializable{
         this.imgUrl = imgUrl;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
